[hotfix][table-planner-blink] Rename utility class and move package
